+++ packages/orq-rc/RELEASES.md
@@ -11039,3 +11039,13 @@ Based on:
 - [python v4.16.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.1 - packages/orq-rc
+
+## 2026-09-18 06:15:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.3 - packages/orq-rc